+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.3] - 2026-03-15
+### Updated
+- Add optional custom list of colors to sector and bars;
+- Add optional custom color for each data item;
+- Add optional custom height and width of chart area;
+- Add optional maximum width of chart area (default is 600px); 
+- Add optional back ground color (default is [surfaceContainerLow] from color scheme);
+
 ## [0.0.2] - 2026-03-04
 ### Updated
 - Add documentation comments